+++ types/js-dom-utils.d.ts
@@ -6,7 +6,6 @@
  * @packageDocumentation
  */
 
-
 /**
  * Focus a DOM element.
  *